Repaired the Slime's shadows and added an image registry for the Slime skins. Also fixed a bug in Time to allow it to start at the exact current time instead of the current hour.

+++ DigIt3/src/com/dig/www/util/Time.java
@@ -28,7 +28,7 @@ public class Time implements ActionListener {
 		String time = String.format("%tr", new Date());
 		String[] s = time.split(":");
 
-		this.time = Integer.parseInt(s[0]) + (Integer.parseInt(s[1]) / 100);
+		this.time = Integer.parseInt(s[0]) + (Float.parseFloat(s[1]) / 100);
 		String t = s[s.length - 1];
 
 		isAM = t.endsWith("AM");
